🚧 Migrate Item and Catalog models to Sequelize and remove legacy SQL models
