+++ src/storage/mysql/Catalog.storage.ts
@@ -1,18 +1,14 @@
 import { CustomError } from "bokchalhandler/dist/CustomError";
-import Catalog from "../../models/sql/Catalog.model";
-import Main from "./Main";
+import Catalog from "../../models/sequelize/Catalog.model";
+import "./Main";
 import CatalogStorageInterface from "../inteface/Catalog.interface";
 
-export default class CatalogStorage extends Main implements CatalogStorageInterface {
-  constructor() {
-    super();
-  }
+export default class CatalogStorage implements CatalogStorageInterface {
+  
 
   async getCatalogs() {
     try {
-      await this.connectToMariaDB();
       const catalogs = await Catalog.findAll();
-      await this.closeMariaDBConnection();
       return catalogs;
     } catch (error) {
       console.error("Error getting catalogs:", error);
@@ -22,13 +18,12 @@ export default class CatalogStorage extends Main implements CatalogStorageInterf
 
   async getItemsByCatalogId(catalogId: number): Promise<any> {
     try {
-      await this.connectToMariaDB();
       const catalog = await Catalog.findByPk(catalogId);
       if (!catalog) {
         throw CustomError.notFound("Catalog not found");
       }
       const items = await catalog.items;
-      await this.closeMariaDBConnection();
+
       return items;
     } catch (error) {
       console.error("Error getting items by catalogId:", error);
